Import .zip function implented

+++ ChartSettings.swift
@@ -3,28 +3,6 @@ import PhotosUI
 import SwiftUI
 import UIKit
 import UniformTypeIdentifiers
-struct ShareSheet: UIViewControllerRepresentable {
-    typealias Callback = (_ activityType: UIActivity.ActivityType?, _ completed: Bool, _ returnedItems: [Any]?, _ error: Error?) -> Void
-
-    let activityItems: [Any]
-    let applicationActivities: [UIActivity]? = nil
-    let excludedActivityTypes: [UIActivity.ActivityType]? = nil
-    let callback: Callback? = nil
-
-    func makeUIViewController(context _: Context) -> UIActivityViewController {
-        let controller = UIActivityViewController(
-            activityItems: activityItems,
-            applicationActivities: applicationActivities
-        )
-        controller.excludedActivityTypes = excludedActivityTypes
-        controller.completionWithItemsHandler = callback
-        return controller
-    }
-
-    func updateUIViewController(_: UIActivityViewController, context _: Context) {
-        // nothing to do here
-    }
-}
 
 struct ImagePicker: UIViewControllerRepresentable {
     @Binding var Image: UIImage?
@@ -80,11 +58,13 @@ struct ChartSettings: View {
     @State private var showingMusicPicker = false
     @State private var zipURL: URL?
     @State private var showingExporter = false
+    @State private var showingImporter = false
     var body: some View {
         List {
             Section(header: Text("File Operation:")) {
                 Button("Import Music...") {
-                    showingMusicPicker = true
+//                    showingMusicPicker = true
+                    showingImporter = true
                 }
 
                 Button("Import Photo...") {
@@ -94,9 +74,12 @@ struct ChartSettings: View {
                     do {
                         try _ = data.saveCache()
                         try self.zipURL = data.exportZip()
-                        
+
                     } catch {}
                     showingExporter = true
+                }
+                Button("Import '.zip' file...") {
+                    showingImporter = true
                 }
                 Button("Save to local storage") {
                     do {
@@ -132,7 +115,7 @@ struct ChartSettings: View {
                     Button("[No copyright]", action: {})
                 }
             }.textCase(nil)
-            Section(header: Text("Variables:l")) {
+            Section(header: Text("Variables:")) {
                 Stepper(value: $data.offsetSecond, in: offsetRange, step: 0.005) {
                     Text("Offset: \(NSString(format: "%.3f", data.offsetSecond))s")
                 }
@@ -174,15 +157,12 @@ struct ChartSettings: View {
 
                     })
                 }
-            }.onChange(of: data.highlightedTicks) { _ in
-                //                dataK.highlightedTicks = data.highlightedTicks
-            }.textCase(nil)
+            }.onChange(of: data.highlightedTicks) { _ in }
+                .textCase(nil)
 
             Section(header: Text("Do not change these:")) {
                 Stepper(value: $data.tickPerBeat,
-                        onEditingChanged: { _ in
-                            //                            dataK.tickPerBeat = data.tickPerBeat
-                        }) {
+                        onEditingChanged: { _ in }) {
                     Text("Tick: \(data.tickPerBeat)")
                 }
             }
@@ -190,59 +170,64 @@ struct ChartSettings: View {
             let configuration = PHPickerConfiguration(photoLibrary: PHPhotoLibrary.shared())
             ImagePicker(Image: $data.imageFile, configuration: configuration, isPresented: $showingImagePicker)
         }
-        .fileImporter(
-            isPresented: $showingMusicPicker,
-            allowedContentTypes: [.audio],
-            allowsMultipleSelection: false
-        ) { result in
+        .fileImporter(isPresented: $showingImporter, allowedContentTypes: [.zip, .mp3], allowsMultipleSelection: false) { result in
             do {
                 guard let selectedFile: URL = try result.get().first else { return }
-                if selectedFile.startAccessingSecurityScopedResource() {
-                    let fm = FileManager.default
-                    let urls = fm.urls(for: .documentDirectory, in: .userDomainMask)
-                    if let url = urls.first {
-                        let dirPath = url.appendingPathComponent("tmp")
-                        if !fm.fileExists(atPath: dirPath.path) {
-                            try fm.createDirectory(at: dirPath, withIntermediateDirectories: true, attributes: nil)
+                if selectedFile.pathExtension == "zip" {
+                    if selectedFile.startAccessingSecurityScopedResource() {
+                        let fm = FileManager.default
+                        let urls = fm.urls(for: .documentDirectory, in: .userDomainMask)
+                        if let url = urls.first {
+                            let fileURL = url.appendingPathComponent("import.zip")
+                            if fm.fileExists(atPath: fileURL.path) {
+                                try fm.removeItem(at: fileURL)
+                            }
+                            try fm.copyItem(at: selectedFile, to: fileURL)
+                            try _ = self.data.importZip()
                         }
-                        let fileURL = dirPath.appendingPathComponent("tmp.mp3")
-                        try fm.removeItem(at: fileURL)
-                        try FileManager.default.copyItem(at: selectedFile, to: fileURL)
-                        print(fileURL)
-                        data.audioFileURL = fileURL
+                    } else {
+                        // Handle denied access
                     }
                 } else {
-                    // Handle denied access
+                    if selectedFile.startAccessingSecurityScopedResource() {
+                        let fm = FileManager.default
+                        let urls = fm.urls(for: .documentDirectory, in: .userDomainMask)
+                        if let url = urls.first {
+                            let dirPath = url.appendingPathComponent("tmp")
+                            if !fm.fileExists(atPath: dirPath.path) {
+                                try fm.createDirectory(at: dirPath, withIntermediateDirectories: true, attributes: nil)
+                            }
+                            let fileURL = dirPath.appendingPathComponent("tmp.mp3")
+                            if fm.fileExists(atPath: fileURL.path) {
+                                try fm.removeItem(at: fileURL)
+                            }
+                            try fm.copyItem(at: selectedFile, to: fileURL)
+                            data.audioFileURL = fileURL
+                        }
+                    } else {}
                 }
+
             } catch {
                 // Handle failure.
+                print(error)
             }
         }
-        .fileExporter(isPresented: $showingExporter, document: URLDocument(data: data), contentType: .zip, onCompletion: {(result) in
-            
-        })
-        // add file export Logic here.
+        .fileExporter(isPresented: $showingExporter, document: URLExportDocument(data: data), contentType: .zip, onCompletion: { _ in })
     }
 }
 
-struct URLDocument: FileDocument{
+struct URLExportDocument: FileDocument {
     static var readableContentTypes: [UTType] = [.zip]
-    
+
     var data: DataStructure?
-    
-    init (data: DataStructure){
-//        self.url = url
+
+    init(data: DataStructure) {
         self.data = data
     }
-    
-    init(configuration: ReadConfiguration) throws {
-        
-    }
-    
-    func fileWrapper(configuration: WriteConfiguration) throws -> FileWrapper {
-//        if self.url != nil{
-//            return try FileWrapper(url: self.url!)
-//        }
+
+    init(configuration _: ReadConfiguration) throws {}
+
+    func fileWrapper(configuration _: WriteConfiguration) throws -> FileWrapper {
         return try FileWrapper(url: try data!.exportZip())
     }
 }
